+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to
 ### Added
 
 - cosmwasm-std: Implement `Not` for `Uint{64,128,256}` ([#1799]).
-- cosmwasm-std: Implement `into_empty` for `QuerierWrapper`, `Deps` and
-  `DepsMut`.
 - cosmwasm-std: Add iterators for `Coins` ([#1806]).
+- cosmwasm-std: Make `abs_diff` const for `Uint{256,512}` and
+  `Int{64,128,256,512}`. It is now const for all integer types.
+- cosmwasm-std: Implement `TryFrom<Decimal256>` for `Decimal` ([#1832])
 
 [#1799]: https://github.com/CosmWasm/cosmwasm/pull/1799
 [#1806]: https://github.com/CosmWasm/cosmwasm/pull/1806
+[#1832]: https://github.com/CosmWasm/cosmwasm/pull/1832
 
 ### Changed
 
@@ -29,6 +31,7 @@ and this project adheres to
 - cosmwasm-check: Update clap dependency to version 4 ([#1677])
 - cosmwasm-vm: Use `wasmparser` for initial validation instead of `parity-wasm`
   ([#1786])
+- cosmwasm-std: Make constructors `Decimal{,256}::{percent,permille,bps}` const
 
 [#1667]: https://github.com/CosmWasm/cosmwasm/pull/1667
 [#1674]: https://github.com/CosmWasm/cosmwasm/pull/1674
@@ -37,6 +40,13 @@ and this project adheres to
 [#1701]: https://github.com/CosmWasm/cosmwasm/pull/1701
 [#1786]: https://github.com/CosmWasm/cosmwasm/pull/1786
 [#1793]: https://github.com/CosmWasm/cosmwasm/pull/1793
+
+## [1.3.3] - 2023-08-22
+
+### Added
+
+- cosmwasm-std: Implement `into_empty` for `QuerierWrapper`, `Deps` and
+  `DepsMut`.
 
 ## [1.3.2] - 2023-08-15
 
@@ -1831,7 +1841,8 @@ Some main points:
 
 All future Changelog entries will reference this base
 
-[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.2...HEAD
+[unreleased]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.3...HEAD
+[1.3.3]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.2...v1.3.3
 [1.3.2]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/CosmWasm/cosmwasm/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/CosmWasm/cosmwasm/compare/v1.2.7...v1.3.0